Split cost pricing tests
